+++ src/main/java/com/github/ompc/greys/command/SessionCommand.java
@@ -10,8 +10,8 @@ import java.lang.instrument.Instrumentation;
 import java.nio.charset.Charset;
 import java.nio.charset.UnsupportedCharsetException;
 
-import static com.github.ompc.greys.util.StringUtil.isNotBlank;
 import static java.lang.String.format;
+import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 /**
  * 查看会话状态命令
