+++ src/main/java/com/github/ompc/greys/server/DefaultCommandHandler.java
@@ -29,9 +29,10 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.logging.Logger;
 
-import static com.github.ompc.greys.util.StringUtil.*;
+import static com.github.ompc.greys.util.GaStringUtils.*;
 import static java.lang.String.format;
 import static java.util.logging.Level.*;
+import static org.apache.commons.lang3.StringUtils.isBlank;
 
 /**
  * 命令处理器
