+++ src/main/java/com/github/ompc/greys/server/Session.java
@@ -1,7 +1,6 @@
 package com.github.ompc.greys.server;
 
 import com.github.ompc.greys.advisor.AdviceWeaver;
-import com.github.ompc.greys.util.IOUtil;
 import com.github.ompc.greys.util.LogUtil;
 
 import java.nio.channels.SocketChannel;
@@ -11,10 +10,13 @@ import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Logger;
 
-import static com.github.ompc.greys.util.StringUtil.*;
+import static com.github.ompc.greys.util.GaStringUtils.DEFAULT_PROMPT;
 import static java.lang.String.format;
 import static java.lang.System.currentTimeMillis;
 import static java.util.logging.Level.INFO;
+import static org.apache.commons.io.IOUtils.closeQuietly;
+import static org.apache.commons.lang3.StringUtils.EMPTY;
+import static org.apache.commons.lang3.StringUtils.isNotBlank;
 
 /**
  * 服务端会话
@@ -76,7 +78,7 @@ public class Session {
     public void destroy() {
 
         isDestroy = true;
-        IOUtil.close(socketChannel);
+        closeQuietly(socketChannel);
 
         if (logger.isLoggable(INFO)) {
             logger.log(INFO, format("session[%d] destroyed.", sessionId));
