+++ src/main/java/com/github/ompc/greys/util/FeatureCodec.java
@@ -5,9 +5,9 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Stack;
 
-import static com.github.ompc.greys.util.CheckUtil.isEquals;
-import static com.github.ompc.greys.util.CheckUtil.isIn;
-import static com.github.ompc.greys.util.StringUtil.isBlank;
+import static com.github.ompc.greys.util.GaCheckUtils.isEquals;
+import static com.github.ompc.greys.util.GaCheckUtils.isIn;
+import static org.apache.commons.lang3.StringUtils.isBlank;
 
 /**
  * Feature编解器(线程安全)<br/>
