--- src/main/java/com/github/ompc/greys/util/StringUtil.java
@@ -1,471 +1,0 @@
-package com.github.ompc.greys.util;
-
-import java.lang.reflect.Modifier;
-import java.util.ArrayList;
-import java.util.Scanner;
-import java.util.regex.Pattern;
-
-/**
- * 字符串工具类
- * Created by vlinux on 15/5/18.
- */
-public class StringUtil {
-
-    /**
-     * 判断是否为空白字符串
-     *
-     * @param string 目标字符串
-     * @return true : 空白字符串 / false : 非空白字符串
-     */
-    public static boolean isBlank(String string) {
-        return null == string
-                || string.length() == 0
-                || string.trim().length() == 0;
-    }
-
-
-    /**
-     * 判断是否为非空白字符串
-     *
-     * @param string 目标字符串
-     * @return true : 非空白字符串 / false : 空白字符串
-     */
-    public static boolean isNotBlank(String string) {
-        return !isBlank(string);
-    }
-
-    /**
-     * 命令提示符
-     */
-    public static final String DEFAULT_PROMPT = "ga?>";
-
-
-    /**
-     * 空字符串
-     */
-    public static final String EMPTY = "";
-
-    /**
-     * 中断提示
-     */
-    public static final String ABORT_MSG = "Press Ctrl+D or Ctrl+X to abort.";
-
-    /**
-     * 拆分参数，要求能将命令行字符串拆分成为多个数组
-     *
-     * @param argumentString 参数行
-     * @return 拆分后的参数数组
-     */
-    public static String[] splitForArgument(String argumentString) {
-
-        Pattern compile = Pattern.compile("(\"[^\"]*\")|('[^']*')|([^\\s+]+)");
-        java.util.regex.Matcher matcher = compile.matcher(argumentString);
-
-        final ArrayList<String> stringList = new ArrayList<String>();
-        while (matcher.find()) {
-
-            final String segment = matcher.group();
-            if (segment.length() > 1
-                    && ((segment.startsWith("'") && segment.endsWith("'")) || (segment.startsWith("\"") && segment.endsWith("\"")))) {
-                stringList.add(segment.replaceAll("(^['|\"])|(['|\"]$)", ""));
-            } else {
-                stringList.add(segment);
-            }
-
-        }
-
-        return stringList.toArray(new String[stringList.size()]);
-
-    }
-
-
-    /**
-     * 获取异常的原因描述
-     *
-     * @param t 异常
-     * @return 异常原因
-     */
-    public static String getCauseMessage(Throwable t) {
-        if (null != t.getCause()) {
-            return getCauseMessage(t.getCause());
-        }
-        return t.getMessage();
-    }
-
-    /**
-     * 展示logo
-     *
-     * @return Logo信息
-     */
-    public static String getLogo() {
-        final StringBuilder logoSB = new StringBuilder();
-        final Scanner scanner = new Scanner(Object.class.getResourceAsStream("/com/github/ompc/greys/res/logo.txt"));
-        while (scanner.hasNextLine()) {
-            logoSB.append(scanner.nextLine()).append("\n");
-        }
-        return logoSB.toString();
-    }
-
-    /**
-     * 将一个对象转换为字符串
-     *
-     * @param obj 目标对象
-     * @return 字符串
-     */
-    public static String newString(Object obj) {
-        if (null == obj) {
-            return EMPTY;
-        }
-        return obj.toString();
-    }
-
-
-    /**
-     * <p>The maximum size to which the padding constant(s) can expand.</p>
-     */
-    private static final int PAD_LIMIT = 8192;
-
-    /**
-     * <p>Repeat a String <code>repeat</code> times to form a
-     * new String.</p>
-     * <p/>
-     * <pre>
-     * StringUtils.repeat(null, 2) = null
-     * StringUtils.repeat("", 0)   = ""
-     * StringUtils.repeat("", 2)   = ""
-     * StringUtils.repeat("a", 3)  = "aaa"
-     * StringUtils.repeat("ab", 2) = "abab"
-     * StringUtils.repeat("a", -2) = ""
-     * </pre>
-     *
-     * @param str    the String to repeat, may be null
-     * @param repeat number of times to repeat str, negative treated as zero
-     * @return a new String consisting of the original String repeated,
-     * <code>null</code> if null String input
-     */
-    public static String repeat(String str, int repeat) {
-        // Performance tuned for 2.0 (JDK1.4)
-
-        if (str == null) {
-            return null;
-        }
-        if (repeat <= 0) {
-            return EMPTY;
-        }
-        int inputLength = str.length();
-        if (repeat == 1 || inputLength == 0) {
-            return str;
-        }
-        if (inputLength == 1 && repeat <= PAD_LIMIT) {
-            return padding(repeat, str.charAt(0));
-        }
-
-        int outputLength = inputLength * repeat;
-        switch (inputLength) {
-            case 1:
-                char ch = str.charAt(0);
-                char[] output1 = new char[outputLength];
-                for (int i = repeat - 1; i >= 0; i--) {
-                    output1[i] = ch;
-                }
-                return new String(output1);
-            case 2:
-                char ch0 = str.charAt(0);
-                char ch1 = str.charAt(1);
-                char[] output2 = new char[outputLength];
-                for (int i = repeat * 2 - 2; i >= 0; i--, i--) {
-                    output2[i] = ch0;
-                    output2[i + 1] = ch1;
-                }
-                return new String(output2);
-            default:
-//                StrBuilder buf = new StrBuilder(outputLength);
-//                for (int i = 0; i < repeat; i++) {
-//                    buf.append(str);
-//                }
-//                return buf.toString();
-
-                final StringBuilder buf = new StringBuilder();
-                for (int i = 0; i < repeat; i++) {
-                    buf.append(str);
-                }
-                return buf.toString();
-        }
-    }
-
-    /**
-     * <p>Returns padding using the specified delimiter repeated
-     * to a given length.</p>
-     * <p/>
-     * <pre>
-     * StringUtils.padding(0, 'e')  = ""
-     * StringUtils.padding(3, 'e')  = "eee"
-     * StringUtils.padding(-2, 'e') = IndexOutOfBoundsException
-     * </pre>
-     * <p/>
-     * <p>Note: this method doesn't not support padding with
-     * <a href="http://www.unicode.org/glossary/#supplementary_character">Unicode Supplementary Characters</a>
-     * as they require a pair of <code>char</code>s to be represented.
-     * If you are needing to support full I18N of your applications
-     * consider using {@link #repeat(String, int)} instead.
-     * </p>
-     *
-     * @param repeat  number of times to repeat delim
-     * @param padChar character to repeat
-     * @return String with repeated character
-     * @throws IndexOutOfBoundsException if <code>repeat &lt; 0</code>
-     * @see #repeat(String, int)
-     */
-    private static String padding(int repeat, char padChar) throws IndexOutOfBoundsException {
-        if (repeat < 0) {
-            throw new IndexOutOfBoundsException("Cannot pad a negative amount: " + repeat);
-        }
-        final char[] buf = new char[repeat];
-        for (int i = 0; i < buf.length; i++) {
-            buf[i] = padChar;
-        }
-        return new String(buf);
-    }
-
-    /**
-     * Gets a String's length or <code>0</code> if the String is <code>null</code>.
-     *
-     * @param str a String or <code>null</code>
-     * @return String length or <code>0</code> if the String is <code>null</code>.
-     * @since 2.4
-     */
-    public static int length(String str) {
-        return str == null ? 0 : str.length();
-    }
-
-
-    /**
-     * 翻译类名称
-     *
-     * @param clazz Java类
-     * @return 翻译值
-     */
-    public static String tranClassName(Class<?> clazz) {
-        if (clazz.isArray()) {
-            StringBuilder sb = new StringBuilder(clazz.getName());
-            sb.delete(0, 2);
-            if (sb.length() > 0 && sb.charAt(sb.length() - 1) == ';') {
-                sb.deleteCharAt(sb.length() - 1);
-            }
-            sb.append("[]");
-            return sb.toString();
-        } else {
-            return clazz.getName();
-        }
-    }
-
-
-    /**
-     * 翻译类名称<br/>
-     * 将 java/lang/String 的名称翻译成 java.lang.String
-     *
-     * @param className 类名称 java/lang/String
-     * @return 翻译后名称 java.lang.String
-     */
-    public static String tranClassName(String className) {
-        return replace(className, "/", ".");
-    }
-
-    /**
-     * 翻译Modifier值
-     *
-     * @param mod modifier
-     * @return 翻译值
-     */
-    public static String tranModifier(int mod) {
-        StringBuilder sb = new StringBuilder();
-        if (Modifier.isAbstract(mod)) {
-            sb.append("abstract,");
-        }
-        if (Modifier.isFinal(mod)) {
-            sb.append("final,");
-        }
-        if (Modifier.isInterface(mod)) {
-            sb.append("interface,");
-        }
-        if (Modifier.isNative(mod)) {
-            sb.append("native,");
-        }
-        if (Modifier.isPrivate(mod)) {
-            sb.append("private,");
-        }
-        if (Modifier.isProtected(mod)) {
-            sb.append("protected,");
-        }
-        if (Modifier.isPublic(mod)) {
-            sb.append("public,");
-        }
-        if (Modifier.isStatic(mod)) {
-            sb.append("static,");
-        }
-        if (Modifier.isStrict(mod)) {
-            sb.append("strict,");
-        }
-        if (Modifier.isSynchronized(mod)) {
-            sb.append("synchronized,");
-        }
-        if (Modifier.isTransient(mod)) {
-            sb.append("transient,");
-        }
-        if (Modifier.isVolatile(mod)) {
-            sb.append("volatile,");
-        }
-        if (sb.length() > 0) {
-            sb.deleteCharAt(sb.length() - 1);
-        }
-        return sb.toString();
-    }
-
-
-    /**
-     * 获取方法执行堆栈信息
-     *
-     * @return 方法堆栈信息
-     */
-    public static String getStack() {
-        final StackTraceElement[] stes = Thread.currentThread().getStackTrace();
-        final StringBuilder stSB = new StringBuilder()
-                .append("Thread Info:").append(Thread.currentThread().getName()).append("\n");
-
-        if (null == stes
-                || stes.length == 0
-                || stes.length == 1) {
-            return stSB.toString();
-        }
-
-        for (int index = 4; index < stes.length; index++) {
-            final StackTraceElement ste = stes[index];
-            stSB.append(index == 2 ? "  " : "    at ")
-                    .append(ste.getClassName()).append(".")
-                    .append(ste.getMethodName())
-                    .append("(").append(ste.getFileName()).append(":").append(ste.getLineNumber()).append(")\n");
-        }
-
-        return stSB.toString();
-    }
-
-
-    /**
-     * <p>Checks if a String is empty ("") or null.</p>
-     * <p/>
-     * <pre>
-     * StringUtils.isEmpty(null)      = true
-     * StringUtils.isEmpty("")        = true
-     * StringUtils.isEmpty(" ")       = false
-     * StringUtils.isEmpty("bob")     = false
-     * StringUtils.isEmpty("  bob  ") = false
-     * </pre>
-     * <p/>
-     * <p>NOTE: This method changed callback Lang version 2.0.
-     * It no longer trims the String.
-     * That functionality is available callback isBlank().</p>
-     *
-     * @param str the String to check, may be null
-     * @return <code>true</code> if the String is empty or null
-     */
-    public static boolean isEmpty(String str) {
-        return str == null || str.length() == 0;
-    }
-
-
-    /**
-     * Represents a failed index search.
-     *
-     * @since 2.1
-     */
-    public static final int INDEX_NOT_FOUND = -1;
-
-    /**
-     * <p>Gets the substring after the last occurrence of a separator.
-     * The separator is not returned.</p>
-     * <p/>
-     * <p>A <code>null</code> string input will return <code>null</code>.
-     * An empty ("") string input will return the empty string.
-     * An empty or <code>null</code> separator will return the empty string if
-     * the input string is not <code>null</code>.</p>
-     * <p/>
-     * <p>If nothing is found, the empty string is returned.</p>
-     * <p/>
-     * <pre>
-     * StringUtils.substringAfterLast(null, *)      = null
-     * StringUtils.substringAfterLast("", *)        = ""
-     * StringUtils.substringAfterLast(*, "")        = ""
-     * StringUtils.substringAfterLast(*, null)      = ""
-     * StringUtils.substringAfterLast("abc", "a")   = "bc"
-     * StringUtils.substringAfterLast("abcba", "b") = "a"
-     * StringUtils.substringAfterLast("abc", "c")   = ""
-     * StringUtils.substringAfterLast("a", "a")     = ""
-     * StringUtils.substringAfterLast("a", "z")     = ""
-     * </pre>
-     *
-     * @param str       the String to get a substring from, may be null
-     * @param separator the String to search for, may be null
-     * @return the substring after the last occurrence of the separator,
-     * <code>null</code> if null String input
-     * @since 2.0
-     */
-    public static String substringAfterLast(String str, String separator) {
-        if (isEmpty(str)) {
-            return str;
-        }
-        if (isEmpty(separator)) {
-            return EMPTY;
-        }
-        int pos = str.lastIndexOf(separator);
-        if (pos == INDEX_NOT_FOUND || pos == (str.length() - separator.length())) {
-            return EMPTY;
-        }
-        return str.substring(pos + separator.length());
-    }
-
-
-    /**
-     * 字符串替换
-     *
-     * @param string 原始字符串
-     * @param oldStr 替换目标
-     * @param newStr 替换成
-     * @return 返回替换后的字符串
-     */
-    public static String replace(String string, String oldStr, String newStr) {
-        if (isEmpty(string)) {
-            return string;
-        }
-        return string.replace(oldStr, newStr);
-    }
-
-
-    /**
-     * 自动换行
-     *
-     * @param string 字符串
-     * @param width  行宽
-     * @return 换行后的字符串
-     */
-    public static String wrap(String string, int width) {
-        final StringBuilder sb = new StringBuilder();
-        final char[] buffer = string.toCharArray();
-        int count = 0;
-        for (char c : buffer) {
-            sb.append(c);
-            if( c == '\n' ) {
-                count = 0;
-            } else if (count++ == width) {
-                sb.append("\n");
-                count = 0;
-            }
-        }
-        return sb.toString();
-    }
-
-    public static void main(String... args) {
-        System.out.println(wrap("12\n34567\n890\n", 5));
-    }
-
-
-}
